refactored composeApp code to cleaner package structure
